[python] extract converter_util.cpp from python_converter.cpp
Move 10 leaf utility methods of python_converter into a new
src/python-frontend/converter/converter_util.cpp: convert_expression_to_code,
create_symbol, create_symbol_id (both overloads), get_location_from_decl,
copy_location_fields_from_decl, type_assertions_enabled, is_coverage_mode,
is_pytest_generation_mode, is_model_file, append_models_from_directory.

Pure code movement — no logic change. First step of an incremental TU split
to make python_converter.cpp (11k LoC) reviewable and parallel-buildable.
